Fix: ShopController owner, member 분리
